update liveness checks
